Planning a more robust way to parse IPA symbols. (only comments)
The current code only covers a few IPA symbols (especially a few diacritics). If I want it to cover more diacritics in a robust way, that is easy to understand I need to define a grammar. I am doing that in the comments. The plan is that once I have sketched out the grammar of how IPA symbols are used, then we will use that to parse sequences of IPA symbols (to extract properties like voicing, place, etc.), and to generate IPA symbols.