Remove extra single-space strings.
These strings containing single spaces were from a time when I used the order of the columns on the IPA table as if it were an integer into a list to tell where the place of articulation of a pulmonic consonant was. This technique led to very convoluted code, that was far more problematic than it was worth.