Remove equal intstance for Place.
Equality over places of articulation wasn't really a standard equality. I instead made a function named equivalentInPlace so that it will be clear that we do not mean a strict equality. The main problem is how we compare phonemes with multiple places of articulation, they are equivalent in place of articulation if they share at least one place of articulation.