Remove unused parameters.
Originally, I had been updating the GUI label, but it did not fully refresh. When text was changed previous text remained at the sides of the new text that replaced it.

I did not find a way to refresh labels reliably, so I decided  to put the header in the output text box instead.